pkp/pkp-lib#7698 consolidate all routing resolving logic in the APIHandler
